requirements_bot.txt
motor
discord.py==1.3.4
romme==0.1.0
discord
beautifulsoup4
Flask
wikipedia
python-dateutil
dblpy
pyyaml
psycopg2-binary
pandas
numpy
asyncpg
aiocache
SQLAlchemy